refactor: extract parseContent service
